[INLONG-4476][Sort][Manager] Remove zk and related classes
