+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/sort/util/FieldInfoUtils.java
@@ -19,6 +19,7 @@ package org.apache.inlong.manager.service.sort.util;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
+import org.apache.inlong.common.enums.MetaField;
 import org.apache.inlong.manager.common.enums.FieldType;
 import org.apache.inlong.manager.common.pojo.sink.SinkField;
 import org.apache.inlong.manager.common.pojo.stream.StreamField;
@@ -39,11 +40,6 @@ import org.apache.inlong.sort.formats.common.TimeFormatInfo;
 import org.apache.inlong.sort.formats.common.TimestampFormatInfo;
 import org.apache.inlong.sort.protocol.FieldInfo;
 import org.apache.inlong.sort.protocol.MetaFieldInfo;
-import org.apache.inlong.sort.protocol.MetaFieldInfo.MetaField;
-import org.apache.inlong.sort.protocol.transformation.FieldMappingRule.FieldMappingUnit;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Util for sort field info.
@@ -77,38 +73,6 @@ public class FieldInfoUtils {
     }
 
     /**
-     * Get field info list.
-     * TODO 1. Support partition field(not need to add index at 0), 2. Add is_metadata field in StreamSinkFieldEntity
-     */
-    public static List<FieldMappingUnit> createFieldInfo(
-            List<StreamField> streamFieldList, List<SinkField> fieldList,
-            List<FieldInfo> sourceFields, List<FieldInfo> sinkFields) {
-
-        // Set source field info list.
-        for (StreamField field : streamFieldList) {
-            FieldInfo sourceField = getFieldInfo(field.getFieldName(), field.getFieldType(),
-                    field.getIsMetaField() == 1, field.getMetaFieldName(), field.getFieldFormat());
-            sourceFields.add(sourceField);
-        }
-
-        List<FieldMappingUnit> mappingUnitList = new ArrayList<>();
-        // Get sink field info list, if the field name equals to build-in field, new a build-in field info
-        for (SinkField field : fieldList) {
-            FieldInfo sinkField = getFieldInfo(field.getFieldName(), field.getFieldType(),
-                    field.getIsMetaField() == 1, field.getMetaFieldName(), field.getFieldFormat());
-            sinkFields.add(sinkField);
-            if (StringUtils.isNotBlank(field.getSourceFieldName())) {
-                FieldInfo sourceField = getFieldInfo(field.getSourceFieldName(),
-                        field.getSourceFieldType(), field.getIsMetaField() == 1,
-                        field.getMetaFieldName(), field.getFieldFormat());
-                mappingUnitList.add(new FieldMappingUnit(sourceField, sinkField));
-            }
-        }
-
-        return mappingUnitList;
-    }
-
-    /**
      * Get field info by the given field name ant type.
      *
      * @apiNote If the field name equals to build-in field, new a build-in field info
@@ -121,27 +85,6 @@ public class FieldInfoUtils {
         } else {
             return new FieldInfo(fieldName, convertFieldFormat(fieldType.toLowerCase(), format));
         }
-    }
-
-    /**
-     * Get all migration field mapping unit list for binlog source.
-     */
-    public static List<FieldMappingUnit> setAllMigrationFieldMapping(List<FieldInfo> sourceFields,
-            List<FieldInfo> sinkFields) {
-        List<FieldMappingUnit> mappingUnitList = new ArrayList<>();
-        MetaFieldInfo dataField = new MetaFieldInfo("data", MetaField.DATA);
-        sourceFields.add(dataField);
-        sinkFields.add(dataField);
-        mappingUnitList.add(new FieldMappingUnit(dataField, dataField));
-        // TODO discarded later
-        for (MetaField metaField : MetaField.values()) {
-            MetaFieldInfo fieldInfo = new MetaFieldInfo(metaField.name(), metaField);
-            sourceFields.add(fieldInfo);
-            sinkFields.add(fieldInfo);
-            mappingUnitList.add(new FieldMappingUnit(fieldInfo, fieldInfo));
-        }
-
-        return mappingUnitList;
     }
 
     /**
