+++ inlong-sort/sort-api/pom.xml
@@ -46,10 +46,6 @@
             <version>${project.version}</version>
         </dependency>
 
-        <dependency>
-            <groupId>org.apache.curator</groupId>
-            <artifactId>curator-framework</artifactId>
-        </dependency>
     </dependencies>
 
 </project>
