+++ inlong-sort/sort-common/src/main/java/org/apache/inlong/sort/protocol/MetaFieldInfo.java
@@ -20,6 +20,7 @@ package org.apache.inlong.sort.protocol;
 import lombok.Getter;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonCreator;
 import org.apache.flink.shaded.jackson2.com.fasterxml.jackson.annotation.JsonProperty;
+import org.apache.inlong.common.enums.MetaField;
 
 /**
  * Meta field info.
@@ -62,75 +63,5 @@ public class MetaFieldInfo extends FieldInfo {
         MetaFieldInfo that = (MetaFieldInfo) o;
         return metaField == that.metaField
                 && super.equals(that);
-    }
-
-    public enum MetaField {
-
-        /**
-         * The process time of flink
-         */
-        PROCESS_TIME,
-        /**
-         * Name of the schema that contain the row, currently used for Oracle database
-         */
-        SCHEMA_NAME,
-        /**
-         * Name of the database that contain the row.
-         */
-        DATABASE_NAME,
-        /**
-         * Name of the table that contain the row.
-         */
-        TABLE_NAME,
-        /**
-         * It indicates the time that the change was made in the database.
-         * If the record is read from snapshot of the table instead of the change stream, the value is always 0
-         */
-        OP_TS,
-        /**
-         * Whether the DDL statement. Currently, it is used for MySQL database.
-         */
-        IS_DDL,
-        /**
-         * Type of database operation, such as INSERT/DELETE, etc. Currently, it is used for MySQL database.
-         */
-        OP_TYPE,
-        /**
-         * MySQL binlog data Row. Currently, it is used for MySQL database.
-         */
-        DATA,
-        /**
-         * The value of the field before update. Currently, it is used for MySQL database.
-         */
-        UPDATE_BEFORE,
-        /**
-         * Batch id of binlog. Currently, it is used for MySQL database.
-         */
-        BATCH_ID,
-        /**
-         * Mapping of sql_type table fields to java data type IDs. Currently, it is used for MySQL database.
-         */
-        SQL_TYPE,
-        /**
-         * The current time when the ROW was received and processed. Currently, it is used for MySQL database.
-         */
-        TS,
-        /**
-         * The table structure. It is only used for MySQL database
-         */
-        MYSQL_TYPE,
-        /**
-         * Primary key field name. Currently, it is used for MySQL database.
-         */
-        PK_NAMES;
-
-        public static MetaField forName(String name) {
-            for (MetaField metaField : values()) {
-                if (metaField.name().equals(name)) {
-                    return metaField;
-                }
-            }
-            throw new UnsupportedOperationException(String.format("Unsupported MetaField=%s for Inlong", name));
-        }
     }
 }
