+++ licenses/inlong-manager/LICENSE
@@ -537,9 +537,6 @@ The text of each license is the standard Apache 2.0 license.
   commons-pool:commons-pool:1.5.4 - Commons Pool (https://commons.apache.org/proper/commons-pool/), (The Apache Software License, Version 2.0)
   org.apache.commons:commons-text:1.9 - Apache Commons Text (https://commons.apache.org/proper/commons-text), (Apache License, Version 2.0)
   com.typesafe:config:1.3.3 - config (https://github.com/lightbend/config/tree/v1.3.3), (Apache License, Version 2.0)
-  org.apache.curator:curator-client:2.13.0 - Curator Client (https://curator.apache.org/curator-client), (The Apache Software License, Version 2.0)
-  org.apache.curator:curator-framework:2.12.0 - Curator Framework (https://curator.apache.org/curator-framework), (The Apache Software License, Version 2.0)
-  org.apache.curator:curator-recipes:2.12.0 - Curator Recipes (https://curator.apache.org/curator-recipes), (The Apache Software License, Version 2.0)
   org.apache.derby:derby:10.14.1.0 - Apache Derby Database Engine and Embedded JDBC Driver (http://db.apache.org/derby/), (Apache 2)
   com.alibaba:druid:1.2.6 -  druid (https://github.com/alibaba/druid/tree/1.2.6), (Apache 2)
   com.alibaba:druid-spring-boot-starter:1.2.6 - druid-spring-boot-starter (https://github.com/alibaba/druid/tree/1.2.6/druid-spring-boot-starter), (Apache 2)
