+++ licenses/inlong-manager/NOTICE
@@ -511,45 +511,6 @@ The Apache Software Foundation (https://www.apache.org/).
 
 ========================================================================
 
-Curator Client NOTICE
-========================================================================
-
-Curator Client
-Copyright 2011-2018 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-
-
-========================================================================
-
-Curator Framework NOTICE
-========================================================================
-
-Curator Framework
-Copyright 2011-2017 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-
-
-========================================================================
-
-Curator Recipes NOTICE
-========================================================================
-
-Curator Recipes
-Copyright 2011-2017 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-
-
-========================================================================
-
 aggs-matrix-stats NOTICE
 ------------------------------------------------------
 lang-mustache NOTICE
