+++ licenses/inlong-sort/LICENSE
@@ -533,9 +533,6 @@ The text of each license is the standard Apache 2.0 license.
   commons-logging:commons-logging:1.2 - Apache Commons Logging (https://commons.apache.org/proper/commons-logging/), (The Apache Software License, Version 2.0)
   commons-net:commons-net:3.1 - Commons Net (https://commons.apache.org/proper/commons-net/), (The Apache Software License, Version 2.0)
   commons-pool:commons-pool:1.5.4 - Commons Pool (https://commons.apache.org/proper/commons-pool/), (The Apache Software License, Version 2.0)
-  org.apache.curator:curator-client:2.12.0 - Curator Client (https://curator.apache.org/curator-client), (The Apache Software License, Version 2.0)
-  org.apache.curator:curator-framework:2.12.0 - Curator Framework (https://curator.apache.org/curator-framework), (The Apache Software License, Version 2.0)
-  org.apache.curator:curator-recipes:2.12.0 - Curator Recipes (https://curator.apache.org/curator-recipes), (The Apache Software License, Version 2.0)
   io.debezium:debezium-api:1.5.4.Final - Debezium API (https://github.com/debezium/debezium/tree/v1.5.4.Final/debezium-api), (Apache License, Version 2.0)
   io.debezium:debezium-connector-mysql:1.5.4.Final - Debezium Connector for MySQL (https://github.com/debezium/debezium/tree/v1.5.4.Final/debezium-connector-mysql), (Apache License, Version 2.0)
   io.debezium:debezium-connector-postgres:1.5.4.Final - Debezium Connector for PostgreSQL (https://github.com/debezium/debezium/tree/v1.5.4.Final/debezium-connector-postgres), (Apache Software License 2.0)
