+++ licenses/inlong-sort/NOTICE
@@ -445,45 +445,6 @@ The Apache Software Foundation (http://www.apache.org/).
 
 ========================================================================
 
-Curator Client NOTICE
-========================================================================
-
-Curator Client
-Copyright 2011-2017 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-
-
-========================================================================
-
-Curator Framework NOTICE
-========================================================================
-
-Curator Framework
-Copyright 2011-2017 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-
-
-========================================================================
-
-Curator Recipes NOTICE
-========================================================================
-
-Curator Recipes
-Copyright 2011-2017 The Apache Software Foundation
-
-This product includes software developed at
-The Apache Software Foundation (http://www.apache.org/).
-
-
-
-========================================================================
-
 Hadoop Metrics2 Reporter for Dropwizard Metrics NOTICE
 ========================================================================
